requirements.txt
Core Agent Framework

crewai==0.35.6
pydantic==2.5.3

LLM Integration (Assuming Google GenAI)

google-genai==0.16.0

General Utilities

python-dotenv